make accordion show referenced item both ways, reference to element, element to reference
